+++ github_action_scripts/python_checks.sh
@@ -10,7 +10,7 @@ python -m venv "${HOME}/venv"
 source "${HOME}/venv/bin/activate"
 python -VV
 python -m site
-python -m pip install -U pip
+python -m pip install -U pip setuptools
 echo dumping pre-installed packages
 python -m pip freeze
 echo installing pip packages